src/ml_consumer/pyproject.toml
[project]
dependencies = [
  "aio-pika>=9.5.5",
  "pyyaml>=6.0.2",
  "opentelemetry-exporter-otlp>=1.33.1",
  "opentelemetry-instrumentation-aio-pika>=0.54b1",
  "opentelemetry-distro>=0.54b1",
  "cassandra-driver>=3.29.2",
  "util",
]

name = "ml-consumer"
version = "0.1.0"
requires-python = ">= 3.10.12"

[tool.uv.sources]
util = { workspace = true }
